fix(sprint11): déplace traits DatabaseInterface + LogicTrait dans mod.rs (avaient été extraits par erreur)
