+++ src/logic/mod.rs
@@ -396,6 +396,64 @@ mod calendar_impl;
 
 
 #[async_trait::async_trait]
+
+#[cfg(test)]
+use mockall::{automock, predicate::*};
+
+#[cfg_attr(test, automock)]
+#[async_trait::async_trait]
+pub trait DatabaseInterface: Send + Sync {
+    async fn insert_user(&self, user: User) -> Result<()>;
+    async fn find_user(&self, username: &str, password: &str) -> Result<Option<User>>;
+    async fn find_emails(&self, mailbox: &str) -> Result<Vec<Email>>;
+    async fn find_email(&self, email_id: &str) -> Result<Option<Email>>;
+    async fn update_email_flag(&self, email_id: &str, flag: &str) -> Result<()>;
+    async fn delete_email(&self, email_id: &str) -> Result<()>;
+    async fn archive_email(&self, email_id: &str) -> Result<()>;
+    async fn select_mailbox(&self, mailbox: &str) -> Result<Mailbox>;
+    async fn search_messages(&self, criteria: &str) -> Result<Vec<u32>>;
+    async fn expunge_mailbox(&self) -> Result<Vec<u32>>;
+    async fn copy_messages(&self, sequence_set: &str, target_mailbox: &str) -> Result<()>;
+    async fn store_flags(&self, sequence_set: &str, flags: Vec<String>, mode: &str) -> Result<()>;
+    async fn find_mailbox(&self, name: &str) -> Result<Option<Mailbox>>;
+    async fn update_mailbox(&self, mailbox: &str, update: Mailbox) -> Result<()>;
+    async fn create_mailbox(&self, mailbox: &str) -> Result<()>;
+    async fn delete_mailbox(&self, mailbox: &str) -> Result<()>;
+    async fn rename_mailbox(&self, old_name: &str, new_name: &str) -> Result<()>;
+    async fn subscribe_mailbox(&self, mailbox: &str) -> Result<()>;
+    async fn unsubscribe_mailbox(&self, mailbox: &str) -> Result<()>;
+    async fn list_subscribed_mailboxes(
+        &self,
+        username: &str,
+        reference: &str,
+        pattern: &str,
+    ) -> Result<Vec<String>>;
+    async fn get_mailbox_status_items(
+        &self,
+        username: &str,
+        mailbox: &str,
+        items: &str,
+    ) -> Result<String>;
+    async fn store_email(&self, username: &str, mailbox: &str, message: &str) -> Result<()>;
+    async fn get_mailbox_status(&self, username: &str, mailbox: &str) -> Result<Mailbox>;
+    async fn noop(&self) -> Result<()>;
+    async fn close_mailbox(&self) -> Result<()>;
+    async fn check_mailbox(&self) -> Result<()>;
+    async fn create_user(&self, username: &str, password: &str, mailbox: &str) -> Result<()>;
+    async fn authenticate_user(&self, username: &str, password: &str) -> Result<Option<User>>;
+    async fn list_mailboxes(
+        &self,
+        username: &str,
+        reference: &str,
+        mailbox: &str,
+    ) -> Result<Vec<String>>;
+}
+
+#[async_trait::async_trait]
+pub trait LogicTrait: Send + Sync {
+    async fn create_user(&self, username: &str, password: &str, mailbox: &str) -> Result<()>;
+}
+
 impl LogicTrait for Logic {
     async fn create_user(&self, username: &str, password: &str, mailbox: &str) -> Result<()> {
         self.create_user(username, password, mailbox).await
